Update DHT-11_FC-37_PCF8523_Sleep_RTC_vers.1.1_010121.ino
For the leaf wetness sensor (FC-37) we might want an analog reading to determine how wet the leaf is, instead of currently, we just read whether the circuit determines that it is wet (1) or it isn't(0) based on the setting of the potentiometer on the wetness sensor. I recommended changes to reflect using an analog reading. I don't exactly understand how the DHT sensor is being read but I  look forward to discussing it with you